home-budget-api/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2
pydantic>=2
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
alembic
email-validator
pytest
httpx
